Create index.js
Create a code for a Component that takes two HTML one text box and one label. Label will
be used to display the images. So it should be like this
If I write “Happy” in the text box the label should show happy face (You can use any image)
If I write “Like” in the text box the label should show Like icon
If I write “sad “ the label should show sad emoji.